Write a little less stuff to disk for each entry in a persistent cache. Incremented the artifact cache format version, as this is a breaking change to the cache layout.

+++ subprojects/core/src/main/java/org/gradle/cache/internal/btree/BTreePersistentIndexedCache.java
@@ -314,7 +314,7 @@ public class BTreePersistentIndexedCache<K, V> implements PersistentIndexedCache
         }
 
         @Override
-        protected int getType() {
+        protected byte getType() {
             return 0x55;
         }
 
@@ -353,7 +353,7 @@ public class BTreePersistentIndexedCache<K, V> implements PersistentIndexedCache
         private IndexRoot root;
 
         @Override
-        protected int getType() {
+        protected byte getType() {
             return 0x77;
         }
 
@@ -662,7 +662,7 @@ public class BTreePersistentIndexedCache<K, V> implements PersistentIndexedCache
         }
 
         @Override
-        protected int getType() {
+        protected byte getType() {
             return 0x33;
         }
